+++ lib/bottom_sheet_search.dart
@@ -1,4 +1,6 @@
 library bottom_sheet_search;
 
+export 'src/bottom_sheet_decoration.dart';
 export 'src/bottom_sheet_search_widget.dart';
-export 'src/list_item.dart';
+export 'src/filter_options.dart';
+export 'src/match_type.dart';