+++ lib/src/match_type.dart
@@ -1,0 +1,5 @@
+enum MatchType {
+  contains,
+  startsWith,
+  exact,
+}